Move array-relate tests to test_array
